UI/UX Improvements. Currency change support
Now the tip and total are shown as the user types their bill, instead of after they interact with a controller. The tip slider is now discretized so only even percentages are allowed (instead of numbers like 15.33%, only 15%, 16%, etc.). The currency displayed is now controlled by a variable in the program, which the user will be able to modify on the upcoming settings page.